Android Login and User Creation v2
