+++ pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
   flutter:
     sdk: flutter
   firebase_auth: 0.5.4
+  validators: ^2.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
